feat(auth): update navigation to dashboard and enhance test configuration
