+++ tests/setup.ts
@@ -8,7 +8,7 @@ beforeAll(async () => {
   // Wait for the web server to start
   const waitForWebServer = async () => {
     const fetchStatusPage = async () => {
-      const response = await fetch('http://localhost:3000')
+      const response = await fetch('http://localhost:3000/api/v1/status/database')
 
       if (!response.ok) {
         console.log('Waiting for web server to start...')
